refactor: Awesome filter in multiple component
style: Improve style for awAction (filter, input)
